+++ front-end/src/utils/GenerateIds.tsx
@@ -1,0 +1,13 @@
+const usedIDs: number[] = [];
+
+export const generateID = () => {
+  let randomID;
+  do {
+    randomID = Math.floor(Math.random() * 100);
+  } while (usedIDs.includes(randomID));
+
+  usedIDs.push(randomID);
+  return randomID;
+};
+
+
